Fix Polling to Task change.
* src/system/ompss/SpawnFunction.cpp (SpawnFunction::spawnFunction) :
Change pool check condition.
(nanos6_stream_spawn_function) : Cast enum class to size_t as it is
required.
src/cluster/messages/MessageTaskNew.cpp (MessageTaskNew::handleMessage)
: Cast enum class when passing to spawn.
* src/system/Bootstrap.cpp (nanos6_preinit) : Remove call to
ClusterManager::postinitialize as before could be too early..
(nanos6_init) : Call ClusterManager::postinitialize.
*
src/executors/workflow/ExecutionWorkflowHost.cpp (HostExecutionStep::start)
: Extend condition to pass the _task to Scheduler::addReadyTask instead
of executing inline.
*
src/executors/threads/TaskFinalizationImplementation.hpp (TaskFinalization::disposeTask)
: Condition decrement of SpawnFunction::_pendingSpawnedFunctions for
polling tasks.
* src/cluster/polling-services/ClusterPollingServices.hpp :
* src/cluster/polling-services/MessageDelivery.hpp : Full refactor.
*
src/cluster/messages/MessageSysFinish.cpp (MessageSysFinish::handleMessage)
: Fix forgotten coma in previous commits.
* src/cluster/ClusterManager.cpp (_activeClusterPollingServices) : New
extern static variable.

Node: This commit was modified during rebase

+++ src/cluster/messages/MessageTaskNew.cpp
@@ -67,7 +67,7 @@ bool MessageTaskNew::handleMessage()
 		TaskOffloading::remoteTaskCleanup, this,
 		"remote-task-wrapper",
 		true,
-		Task::nanos6_task_runtime_flag_t::nanos6_remote_wrapper_flag
+		(size_t) Task::nanos6_task_runtime_flag_t::nanos6_remote_wrapper_flag
 	);
 
 	// The Message will be deleted by remoteTaskCleanup
